Don't sleep to "wait for the chunk generator to catch up"
Not sure what the logic behind this was. Chunk generation currently
happens synchronously, and chunk loads caused by the API are
synchronous. So if we sleep here, we aren't accomplishing anything since
we're sleeping the main thread anyways and the aforementioned things
we're "waiting" for occur on the main thread (although the story may be
different in Paper). Maybe this was async once upon a time?